# Improve set operation speeds