--- a/sparse.go
+++ b/sparse.go
@@ -2,6 +2,8 @@ package hyperloglog
 
 import (
 	"math/bits"
+
+	"github.com/kamstrup/intmap"
 )
 
 func getIndex(k uint32, p, pp uint8) uint32 {
@@ -34,37 +36,61 @@ func decodeHash(k uint32, p, pp uint8) (uint32, uint8) {
 	return getIndex(k, p, pp), r
 }
 
-type set map[uint32]struct{}
+type set struct {
+	m *intmap.Set[uint32]
+}
 
-func (s set) add(v uint32) bool {
-	_, ok := s[v]
-	if ok {
+func newSet(size int) *set {
+	return &set{m: intmap.NewSet[uint32](size)}
+}
+
+func (s *set) ForEach(fn func(v uint32)) {
+	s.m.ForEach(func(v uint32) bool {
+		fn(v)
+		return true
+	})
+}
+
+func (s *set) Merge(other *set) {
+	other.m.ForEach(func(v uint32) bool {
+		s.m.Add(v)
+		return true
+	})
+}
+
+func (s *set) Len() int {
+	return s.m.Len()
+}
+
+func (s *set) add(v uint32) bool {
+	if s.m.Has(v) {
 		return false
 	}
-	s[v] = struct{}{}
+	s.m.Add(v)
 	return true
 }
 
-func (s set) Clone() set {
+func (s *set) Clone() *set {
 	if s == nil {
 		return nil
 	}
 
-	newS := make(map[uint32]struct{}, len(s))
-	for k, v := range s {
-		newS[k] = v
-	}
-	return newS
+	newS := intmap.NewSet[uint32](s.m.Len())
+	s.m.ForEach(func(v uint32) bool {
+		newS.Add(v)
+		return true
+	})
+	return &set{m: newS}
 }
 
-func (s set) MarshalBinary() (data []byte, err error) {
+func (s *set) MarshalBinary() (data []byte, err error) {
 	// 4 bytes for the size of the set, and 4 bytes for each key.
 	// list.
-	data = make([]byte, 0, 4+(4*len(s)))
+	data = make([]byte, 0, 4+(4*s.m.Len()))
 
 	// Length of the set. We only need 32 bits because the size of the set
 	// couldn't exceed that on 32 bit architectures.
-	sl := len(s)
+	sl := s.m.Len()
 	data = append(data, []byte{
 		byte(sl >> 24),
 		byte(sl >> 16),
@@ -73,14 +99,15 @@ func (s set) MarshalBinary() (data []byte, err error) {
 	}...)
 
 	// Marshal each element in the set.
-	for k := range s {
+	s.m.ForEach(func(k uint32) bool {
 		data = append(data, []byte{
 			byte(k >> 24),
 			byte(k >> 16),
 			byte(k >> 8),
 			byte(k),
 		}...)
-	}
+		return true
+	})
 
 	return data, nil
 }