use intmap.Set instead of map[uint32]struct for set

```
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/axiomhq/hyperloglog
cpu: Apple M3 Max
                                      │    old.txt    │               new.txt               │
                                      │    sec/op     │    sec/op     vs base               │
_Merge/size1=100/size2=100-16           17.976µ ± ∞ ¹   7.794µ ± ∞ ¹  -56.64% (p=0.008 n=5)
_Merge/size1=100/size2=10000-16          31.91µ ± ∞ ¹   26.78µ ± ∞ ¹  -16.08% (p=0.008 n=5)
_Merge/size1=100/size2=1000000-16        16.70µ ± ∞ ¹   11.86µ ± ∞ ¹  -29.00% (p=0.008 n=5)
_Merge/size1=10000/size2=100-16          22.21µ ± ∞ ¹   23.00µ ± ∞ ¹        ~ (p=0.548 n=5)
_Merge/size1=10000/size2=10000-16        61.63µ ± ∞ ¹   60.64µ ± ∞ ¹   -1.60% (p=0.008 n=5)
_Merge/size1=10000/size2=1000000-16      34.78µ ± ∞ ¹   33.31µ ± ∞ ¹        ~ (p=0.222 n=5)
_Merge/size1=1000000/size2=100-16        8.316µ ± ∞ ¹   7.948µ ± ∞ ¹   -4.43% (p=0.008 n=5)
_Merge/size1=1000000/size2=10000-16      14.68µ ± ∞ ¹   14.38µ ± ∞ ¹   -2.05% (p=0.008 n=5)
_Merge/size1=1000000/size2=1000000-16    29.66µ ± ∞ ¹   29.03µ ± ∞ ¹        ~ (p=0.548 n=5)
geomean                                  22.78µ         19.36µ        -15.02%
¹ need >= 6 samples for confidence interval at level 0.95

                                      │    old.txt    │                new.txt                │
                                      │     B/op      │      B/op       vs base               │
_Merge/size1=100/size2=100-16           5.847Ki ± ∞ ¹   10.008Ki ± ∞ ¹  +71.17% (p=0.008 n=5)
_Merge/size1=100/size2=10000-16         18.96Ki ± ∞ ¹    21.07Ki ± ∞ ¹  +11.16% (p=0.008 n=5)
_Merge/size1=100/size2=1000000-16       18.96Ki ± ∞ ¹    21.07Ki ± ∞ ¹  +11.16% (p=0.008 n=5)
_Merge/size1=10000/size2=100-16         16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
_Merge/size1=10000/size2=10000-16       16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
_Merge/size1=10000/size2=1000000-16     16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
_Merge/size1=1000000/size2=100-16       16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
_Merge/size1=1000000/size2=10000-16     16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
_Merge/size1=1000000/size2=1000000-16   16.14Ki ± ∞ ¹    16.16Ki ± ∞ ¹   +0.15% (p=0.008 n=5)
geomean                                 14.94Ki          16.26Ki         +8.78%
¹ need >= 6 samples for confidence interval at level 0.95

                                      │   old.txt   │              new.txt               │
                                      │  allocs/op  │  allocs/op   vs base               │
_Merge/size1=100/size2=100-16           32.00 ± ∞ ¹   20.00 ± ∞ ¹  -37.50% (p=0.008 n=5)
_Merge/size1=100/size2=10000-16         26.00 ± ∞ ¹   20.00 ± ∞ ¹  -23.08% (p=0.008 n=5)
_Merge/size1=100/size2=1000000-16       26.00 ± ∞ ¹   20.00 ± ∞ ¹  -23.08% (p=0.008 n=5)
_Merge/size1=10000/size2=100-16         4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
_Merge/size1=10000/size2=10000-16       4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
_Merge/size1=10000/size2=1000000-16     4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
_Merge/size1=1000000/size2=100-16       4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
_Merge/size1=1000000/size2=10000-16     4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
_Merge/size1=1000000/size2=1000000-16   4.000 ± ∞ ¹   6.000 ± ∞ ¹  +50.00% (p=0.008 n=5)
geomean                                 7.639         8.963        +17.33%
¹ need >= 6 samples for confidence interval at level 0.95
```