@ilyam8 2024-12-08
@seiflotfy, hey. I had to revert to v0.2.0 because this PR breaks building on 32bit systems because of `github.com/kamstrup/intmap`

```
/home/cm/go/pkg/mod/github.com/kamstrup/intmap@v0.5.0/map64.go:22:11: 0x9E3779B9 (untyped int constant 2654435769) overflows int
```